+++ src/ts/toolbar/Fullscreen.ts
@@ -16,6 +16,7 @@ export class Fullscreen extends MenuItem {
                     this.innerHTML = menuItem.icon;
                 }
                 vditor.element.style.zIndex = "";
+                document.body.style.overflow = "";
                 vditor.element.classList.remove("vditor--fullscreen");
                 Object.keys(vditor.toolbar.elements).forEach((key) => {
                     const svgElement = vditor.toolbar.elements[key].firstChild as HTMLElement;
@@ -31,6 +32,7 @@ export class Fullscreen extends MenuItem {
                     this.innerHTML = '<svg><use xlink:href="#vditor-icon-contract"></use></svg>';
                 }
                 vditor.element.style.zIndex = vditor.options.fullscreen.index.toString();
+                document.body.style.overflow = "hidden";
                 vditor.element.classList.add("vditor--fullscreen");
                 Object.keys(vditor.toolbar.elements).forEach((key) => {
                     const svgElement = vditor.toolbar.elements[key].firstChild as HTMLElement;