+++ __test__/util/Options.test.ts
@@ -102,7 +102,6 @@ describe("Options", () => {
             },
             markdown: {
                 autoSpace: false,
-                chinesePunct: false,
                 codeBlockPreview: true,
                 fixTermTypo: false,
                 footnotes: true,